requirements.txt
datasets 
python-dotenv
tqdm 
chromadb 
langchain-openai 
langchain-community